# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/config": {
      "post": {
        "operationId": "ConfigController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConfigDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建配置",
        "tags": [
          "系统配置"
        ]
      },
      "get": {
        "operationId": "ConfigController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取所有配置",
        "tags": [
          "系统配置"
        ]
      },
      "patch": {
        "operationId": "ConfigController_updateAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新所有配置",
        "tags": [
          "系统配置"
        ]
      }
    },
    "/config/group/{group}": {
      "get": {
        "operationId": "ConfigController_findByGroup",
        "parameters": [
          {
            "name": "group",
            "required": true,
            "in": "path",
            "description": "配置分组",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "根据分组获取配置",
        "tags": [
          "系统配置"
        ]
      },
      "patch": {
        "operationId": "ConfigController_updateGroup",
        "parameters": [
          {
            "name": "group",
            "required": true,
            "in": "path",
            "description": "配置分组",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新分组配置",
        "tags": [
          "系统配置"
        ]
      }
    },
...
    "/collection": {
      "post": {
        "operationId": "CollectionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建合集",
        "tags": [
          "合集管理"
        ]
      },
      "get": {
        "operationId": "CollectionController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "用户ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键词搜索（收藏夹名称）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "排序方向",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "获取合集列表",
        "tags": [
          "合集管理"
        ]
      }
    },
    "/collection/{id}": {
      "get": {
        "operationId": "CollectionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "合集ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "获取合集详情",
        "tags": [
          "合集管理"
        ]
      },
      "patch": {
        "operationId": "CollectionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "合集ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新合集",
        "tags": [
          "合集管理"
        ]
      },
      "delete": {
        "operationId": "CollectionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "合集ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除合集",
        "tags": [
          "合集管理"
        ]
      }
    },
    "/collection/{id}/article/{articleId}": {
      "post": {
        "operationId": "CollectionController_addToCollection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "合集ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "articleId",
            "required": true,
            "in": "path",
            "description": "文章ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "添加文章到合集",
        "tags": [
          "合集管理"
        ]
      },
      "delete": {
        "operationId": "CollectionController_removeFromCollection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "合集ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "articleId",
            "required": true,
            "in": "path",
            "description": "文章ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "从合集移除文章",
        "tags": [
          "合集管理"
        ]
      }
    },
    "/collection/{id}/items": {
      "get": {
        "operationId": "CollectionController_getCollectionItems",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "合集ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "获取合集中的文章列表",
        "tags": [
          "合集管理"
        ]
      }
    },
    "/collection/check/{articleId}": {
      "get": {
        "operationId": "CollectionController_checkArticleInCollections",
        "parameters": [
          {
            "name": "articleId",
            "required": true,
            "in": "path",
            "description": "文章ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "检查文章是否在合集中",
        "tags": [
          "合集管理"
        ]
      }
    },
    "/collection/article/{articleId}/info": {
      "get": {
        "operationId": "CollectionController_getArticleCollectionInfo",
        "parameters": [
          {
            "name": "articleId",
            "required": true,
            "in": "path",
            "description": "文章ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取文章所在的合集信息",
        "tags": [
          "合集管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "PicArt API",
    "description": "PicArt 图片社区 API 文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateConfigDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateCollectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "合集名称",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "合集描述"
          },
          "avatar": {
            "type": "string",
            "description": "头像"
          },
          "cover": {
            "type": "string",
            "description": "封面图片"
          },
          "isPublic": {
            "type": "boolean",
            "description": "是否公开",
            "default": false
          },
          "sort": {
            "type": "number",
            "description": "排序",
            "default": 0
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCollectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "合集名称",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "合集描述"
          },
          "avatar": {
            "type": "string",
            "description": "头像"
          },
          "cover": {
            "type": "string",
            "description": "封面图片"
          },
          "isPublic": {
            "type": "boolean",
            "description": "是否公开",
            "default": false
          },
          "sort": {
            "type": "number",
            "description": "排序",
            "default": 0
          }
        }
      },
...
    }
  }
}